Time: 0 ms (100%), Space: 8.4 MB (93.45%) - LeetHub

+++ 0050-powx-n/0050-powx-n.cpp
@@ -17,7 +17,7 @@ public:
         }
 
         if (n<0){
-            return ans=1/(double)ans;
+            return ans=1/ans;
         }
         return ans;
     }
